Make revision link more discoverable and less clicky
Previously, the entire header was a giant link. This removes the
link functionality. I've added back an icon that links to the revision
but I didn't do something similar for the user. We can add that back
in a future patch.

I expect there to be opinions about the icon and its placement.

Bug: T194439